requirements.txt
kafka-python==2.2.13
setuptools==78.1.1
wheel==0.45.1
